Changes requested via Zi and test fixes
