--- src/utils/types.ts
@@ -1,2 +1,0 @@
-// Type for the emoji reaction options
-export type Emoji = '🔥' | '⚠️' | '🛡' | '💩' | '🎉';
